pyproject.toml
[project]
name = "django-copilotkit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ag-ui-langgraph>=0.0.17",
    "ag-ui-protocol>=0.1.9",
    "django>=5.2.7",
    "django-choices-field>=3.1.0",
    "django-svelte-jsoneditor>=0.4.4",
    "fastmcp>=2.12.4",
    "google-auth>=2.41.1",
    "langchain-deepseek>=0.1.4",
    "langchain-mcp-adapters>=0.1.11",
    "langchain-openai>=0.3.35",
    "langchain-tavily>=0.2.12",
    "mcp[cli]>=1.17.0",
    "openai>=2.3.0",
    "python-decouple>=3.8",
    "python-dotenv>=1.1.1",
    "redis[hiredis]>=6.4.0",
    "shortuuid>=1.0.13",
    "strawberry-graphql-django>=0.66.2",
    "uvicorn>=0.37.0",
]
